updated state manager to be internal to router
